--- packages/qvac-lib-infer-onnx-base/cmake/qvac-lib-infer-onnx-baseConfig.cmake.in
@@ -1,7 +1,0 @@
-@PACKAGE_INIT@
-
-include(CMakeFindDependencyMacro)
-find_dependency(onnxruntime CONFIG)
-
-include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
-check_required_components("@PROJECT_NAME@")
