+++ packages/qvac-lib-infer-onnx/tests/fixtures/identity_int64.onnx
