+++ packages/qvac-lib-infer-onnx/tests/fixtures/multi_output.onnx
