+++ packages/qvac-lib-infer-onnx/vcpkg-configuration.json
